# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an stripe payment example",
        "title": "Stripe Payment Sample",
        "contact": {
            "name": "API Support",
            "email": "dev@mattiamueggler.ch"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/create-subscription": {
            "post": {
                "description": "Create a Subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "description": "PaymentRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessPayment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/pay": {
            "post": {
                "description": "Create a payment without customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "parameters": [
                    {
                        "description": "PaymentRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessPayment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/pay-subscription": {
            "post": {
                "description": "Pay a Subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "parameters": [
                    {
                        "description": "PaymentRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessPayment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/pay-with-customer": {
            "post": {
                "description": "Create a payment with customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "parameters": [
                    {
                        "description": "PaymentRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PaymentRequestWithCustomer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessPayment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "List all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/stripe.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreditCard": {
            "type": "object",
            "properties": {
                "cardnumber": {
                    "type": "string"
                },
                "cvc": {
                    "type": "string"
                },
                "expire_month": {
                    "type": "integer"
                },
                "expire_year": {
                    "type": "integer"
                }
            }
        },
        "main.Customer": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "post_cod": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "street_nr": {
                    "type": "string"
                }
            }
        },
        "main.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.PaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "credit_card": {
                    "$ref": "#/definitions/main.CreditCard"
                },
                "currency": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "main.PaymentRequestWithCustomer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "credit_card": {
                    "$ref": "#/definitions/main.CreditCard"
                },
                "currency": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/main.Customer"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "main.SuccessPayment": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "stripe.Price": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Whether the price can be used for new purchases.",
                    "type": "boolean"
                },
                "billing_scheme": {
                    "description": "Describes how to compute the price per period. Either `per_unit` or `tiered`. `per_unit` indicates that the fixed amount (specified in `unit_amount` or `unit_amount_decimal`) will be charged per unit in `quantity` (for prices with `usage_type=licensed`), or per unit of total usage (for prices with `usage_type=metered`). `tiered` indicates that the unit pricing will be computed using a tiering strategy as defined using the `tiers` and `tiers_mode` attributes.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceBillingScheme"
                        }
                    ]
                },
                "created": {
                    "description": "Time at which the object was created. Measured in seconds since the Unix epoch.",
                    "type": "integer"
                },
                "currency": {
                    "description": "Three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html), in lowercase. Must be a [supported currency](https://stripe.com/docs/currencies).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.Currency"
                        }
                    ]
                },
                "currency_options": {
                    "description": "Prices defined in each available currency option. Each key must be a three-letter [ISO currency code](https://www.iso.org/iso-4217-currency-codes.html) and a [supported currency](https://stripe.com/docs/currencies).",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/stripe.PriceCurrencyOptions"
                    }
                },
                "custom_unit_amount": {
                    "description": "When set, provides configuration for the amount to be adjusted by the customer during Checkout Sessions and Payment Links.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceCustomUnitAmount"
                        }
                    ]
                },
                "deleted": {
                    "type": "boolean"
                },
                "id": {
                    "description": "Unique identifier for the object.",
                    "type": "string"
                },
                "livemode": {
                    "description": "Has the value `true` if the object exists in live mode or the value `false` if the object exists in test mode.",
                    "type": "boolean"
                },
                "lookup_key": {
                    "description": "A lookup key used to retrieve prices dynamically from a static string. This may be up to 200 characters.",
                    "type": "string"
                },
                "metadata": {
                    "description": "Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "nickname": {
                    "description": "A brief description of the price, hidden from customers.",
                    "type": "string"
                },
                "object": {
                    "description": "String representing the object's type. Objects of the same type share the same value.",
                    "type": "string"
                },
                "product": {
                    "description": "The ID of the product this price is associated with.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.Product"
                        }
                    ]
                },
                "recurring": {
                    "description": "The recurring components of a price such as `interval` and `usage_type`.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceRecurring"
                        }
                    ]
                },
                "tax_behavior": {
                    "description": "Specifies whether the price is considered inclusive of taxes or exclusive of taxes. One of `inclusive`, `exclusive`, or `unspecified`. Once specified as either `inclusive` or `exclusive`, it cannot be changed.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceTaxBehavior"
                        }
                    ]
                },
                "tiers": {
                    "description": "Each element represents a pricing tier. This parameter requires `billing_scheme` to be set to `tiered`. See also the documentation for `billing_scheme`.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/stripe.PriceTier"
                    }
                },
                "tiers_mode": {
                    "description": "Defines if the tiering price should be `graduated` or `volume` based. In `volume`-based tiering, the maximum quantity within a period determines the per unit price. In `graduated` tiering, pricing can change as the quantity grows.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceTiersMode"
                        }
                    ]
                },
                "transform_quantity": {
                    "description": "Apply a transformation to the reported usage or set quantity before computing the amount billed. Cannot be combined with `tiers`.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceTransformQuantity"
                        }
                    ]
                },
                "type": {
                    "description": "One of `one_time` or `recurring` depending on whether the price is for a one-time purchase or a recurring (subscription) purchase.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.PriceType"
                        }
                    ]
                },
                "unit_amount": {
                    "description": "The unit amount in %s to be charged, represented as a whole integer if possible. Only set if `billing_scheme=per_unit`.",
                    "type": "integer"
                },
                "unit_amount_decimal": {
                    "description": "The unit amount in %s to be charged, represented as a decimal string with at most 12 decimal places. Only set if `billing_scheme=per_unit`.",
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "stripe.Product": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Whether the product is currently available for purchase.",
                    "type": "boolean"
                },
                "attributes": {
                    "description": "A list of up to 5 attributes that each SKU can provide values for (e.g., `[\"color\", \"size\"]`).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "caption": {
                    "description": "A short one-line description of the product, meant to be displayable to the customer. Only applicable to products of `type=good`.",
                    "type": "string"
                },
                "created": {
                    "description": "Time at which the object was created. Measured in seconds since the Unix epoch.",
                    "type": "integer"
                },
                "deactivate_on": {
                    "description": "An array of connect application identifiers that cannot purchase this product. Only applicable to products of `type=good`.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "default_price": {
                    "description": "The ID of the [Price](https://stripe.com/docs/api/prices) object that is the default price for this product.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.Price"
                        }
                    ]
                },
                "deleted": {
                    "type": "boolean"
                },
                "description": {
                    "description": "The product's description, meant to be displayable to the customer. Use this field to optionally store a long form explanation of the product being sold for your own rendering purposes.",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the object.",
                    "type": "string"
                },
                "images": {
                    "description": "A list of up to 8 URLs of images for this product, meant to be displayable to the customer.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "livemode": {
                    "description": "Has the value `true` if the object exists in live mode or the value `false` if the object exists in test mode.",
                    "type": "boolean"
                },
                "metadata": {
                    "description": "Set of [key-value pairs](https://stripe.com/docs/api/metadata) that you can attach to an object. This can be useful for storing additional information about the object in a structured format.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "The product's name, meant to be displayable to the customer.",
                    "type": "string"
                },
                "object": {
                    "description": "String representing the object's type. Objects of the same type share the same value.",
                    "type": "string"
                },
                "package_dimensions": {
                    "description": "The dimensions of this product for shipping purposes.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.ProductPackageDimensions"
                        }
                    ]
                },
                "shippable": {
                    "description": "Whether this product is shipped (i.e., physical goods).",
                    "type": "boolean"
                },
                "statement_descriptor": {
                    "description": "Extra information about a product which will appear on your customer's credit card statement. In the case that multiple products are billed at once, the first statement descriptor will be used.",
                    "type": "string"
                },
                "tax_code": {
                    "description": "A [tax code](https://stripe.com/docs/tax/tax-categories) ID.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.TaxCode"
                        }
                    ]
                },
                "type": {
                    "description": "The type of the product. The product is either of type `good`, which is eligible for use with Orders and SKUs, or `service`, which is eligible for use with Subscriptions and Plans.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stripe.ProductType"
                        }
                    ]
                },
                "unit_label": {
                    "description": "A label that represents units of this product. When set, this will be included in customers' receipts, invoices, Checkout, and the customer portal.",
                    "type": "string"
                },
                "updated": {
                    "description": "Time at which the object was last updated. Measured in seconds since the Unix epoch.",
                    "type": "integer"
                },
                "url": {
                    "description": "A URL of a publicly-accessible webpage for this product.",
                    "type": "string"
                }
            }
        },
        "stripe.ProductPackageDimensions": {
            "type": "object",
            "properties": {
                "height": {
                    "description": "Height, in inches.",
                    "type": "number"
                },
                "length": {
                    "description": "Length, in inches.",
                    "type": "number"
                },
                "weight": {
                    "description": "Weight, in ounces.",
                    "type": "number"
                },
                "width": {
                    "description": "Width, in inches.",
                    "type": "number"
                }
            }
        },
...
        "stripe.TaxCode": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "A detailed description of which types of products the tax code represents.",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier for the object.",
                    "type": "string"
                },
                "name": {
                    "description": "A short name for the tax code.",
                    "type": "string"
                },
                "object": {
                    "description": "String representing the object's type. Objects of the same type share the same value.",
                    "type": "string"
                }
            }
        }
    }
}